:ok_hand: [#1183] -- processed PR feedback
* Split tests into separate modules
* Mock get_solo call rather than storing the config in DB
